refactor: Players 생성자 매개변수 수정
