+++ src/main/java/controller/EntityCreator.java
@@ -2,9 +2,12 @@ package controller;
 
 import domain.Cards;
 import domain.Dealer;
+import domain.Player;
+import domain.PlayerName;
 import domain.Players;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import static view.InputView.printErrorMessage;
 import static view.InputView.readPlayerNames;
@@ -20,11 +23,21 @@ public class EntityCreator {
 
     private Players getValidPlayerNames() {
         try {
-            return new Players(readPlayerNames());
+            List<String> playerNames = readPlayerNames();
+            List<Player> player = getPlayerList(playerNames);
+            return new Players(player);
         } catch (RuntimeException exception) {
             printErrorMessage(exception);
             return getValidPlayerNames();
         }
+    }
+
+    private List<Player> getPlayerList(List<String> list) {
+        List<Player> player = new ArrayList<>();
+        for (String playerName : list) {
+            player.add(new Player(new PlayerName(playerName), new Cards(new ArrayList<>())));
+        }
+        return player;
     }
 
     public Dealer getDealer() {
